requirements-note.txt
actionlib==1.14.0
angles==1.9.13
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.7
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cupy-cuda11x==12.2.0
cv-bridge==1.16.2
cycler==0.11.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
fastrlock==0.8.1
fonttools==4.39.3
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
idna==3.4
image-geometry==1.16.2
importlib-resources==5.12.0
interactive-markers==1.12.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.4
laser_geometry==1.6.7
matplotlib==3.7.1
message-filters==1.16.0
numpy==1.24.1
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
Pillow==9.3.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2023.3
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.28.1
resource_retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosserial_arduino==0.9.2
rosserial_client==0.9.2
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rtabmap-ros==0.20.23
rviz==1.14.20
scipy==1.10.1
seaborn==0.12.2
sensor-msgs==1.13.1
sentry-sdk==1.20.0
six==1.16.0
smach==2.5.1
smach-ros==2.5.1
smclib==1.8.6
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
thop==0.1.1.post2209072238
topic-tools==1.16.0
torch==1.12.1+cu102
torchaudio==0.12.1+cu102
torchvision==0.13.1+cu102
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.83
urllib3==1.26.13
xacro==1.14.16
zipp==3.15.0
